# secp256k1: Use specialized types in public key.