--- a/dcrec/secp256k1/ciphering.go
+++ b/dcrec/secp256k1/ciphering.go
@@ -45,10 +45,12 @@ var (
 // public key using Diffie-Hellman key exchange (ECDH) (RFC 4753).
 // RFC5903 Section 9 states we should only return x.
 func GenerateSharedSecret(privkey *PrivateKey, pubkey *PublicKey) []byte {
-	privKeyBytes := privkey.Key.Bytes()
-	x, _ := S256().ScalarMult(pubkey.x, pubkey.y, privKeyBytes[:])
-	zeroArray32(&privKeyBytes)
-	return x.Bytes()
+	var point, result JacobianPoint
+	pubkey.AsJacobian(&point)
+	ScalarMultNonConst(&privkey.Key, &point, &result)
+	result.ToAffine()
+	xBytes := result.X.Bytes()
+	return xBytes[:]
 }
 
 // Encrypt encrypts data for the target public key using AES-256-CBC. It also