--- a/dcrec/secp256k1/curve_test.go
+++ b/dcrec/secp256k1/curve_test.go
@@ -736,20 +736,20 @@ func TestSplitKRand(t *testing.T) {
 }
 
 // Test this curve's usage with the ecdsa package.
-func testKeyGeneration(t *testing.T, c *KoblitzCurve, tag string) {
+func testKeyGeneration(t *testing.T, tag string) {
 	priv, err := GeneratePrivateKey()
 	if err != nil {
 		t.Errorf("%s: error: %s", tag, err)
 		return
 	}
 	pub := priv.PubKey()
-	if !c.IsOnCurve(pub.x, pub.y) {
+	if !isOnCurve(&pub.x, &pub.y) {
 		t.Errorf("%s: public key invalid: %s", tag, err)
 	}
 }
 
 func TestKeyGeneration(t *testing.T) {
-	testKeyGeneration(t, S256(), "S256")
+	testKeyGeneration(t, "S256")
 }
 
 func TestNAF(t *testing.T) {