--- a/dcrec/secp256k1/schnorr/signature.go
+++ b/dcrec/secp256k1/schnorr/signature.go
@@ -52,9 +52,8 @@ func NewSignature(r *secp256k1.FieldVal, s *secp256k1.ModNScalar) *Signature {
 // Serialize returns the Schnorr signature in the more strict format.
 //
 // The signatures are encoded as
-//   sig[0:32]  R, a point encoded as big endian
-//   sig[32:64] S, scalar multiplication/addition results = (ab+c) mod l
-//     encoded also as big endian
+//   sig[0:32]  x coordinate of the point R, encoded as a big-endian uint256
+//   sig[32:64] s, encoded also as big-endian uint256
 func (sig Signature) Serialize() []byte {
 	// Total length of returned signature is the length of r and s.
 	var b [SignatureSize]byte
@@ -144,8 +143,7 @@ func schnorrVerify(sig *Signature, hash []byte, pubKey *secp256k1.PublicKey) err
 	// Step 2.
 	//
 	// Fail if Q is not a point on the curve
-	curve := secp256k1.S256()
-	if !curve.IsOnCurve(pubKey.X(), pubKey.Y()) {
+	if !pubKey.IsOnCurve() {
 		str := "pubkey point is not on curve"
 		return signatureError(ErrPubKeyNotOnCurve, str)
 	}