**This requires #2157 and  #2162**.

This updates the internal implementation of the `PublicKey` type to use the `FieldVal` type instead of `big.Int`s and modifies `NewPublicKey` to accept the `FieldVal`s and updates all callers and tests in the repository accordingly.

Public key parsing and serialization are now completely free of `big.Int`s.

Finally, it moves the `X` and `Y` methods to the `ellipticadaptor.go` file since they are only intended to provide interop with the standard library elliptic curve adaptor code.

This reduces the number of allocations needed in the ecdsa compact signature recovery as well as schnorr signature verification and results in a slight speedup for both.

The following benchmarks show the before and after comparison:

```
benchmark                 old ns/op    new ns/op    delta
----------------------------------------------------------
BenchmarkRecoverCompact   198012       196173       -0.93%
BenchmarkSigVerify        179170       177805       -0.76%

benchmark                 old allocs   new allocs   delta
-----------------------------------------------------------
BenchmarkRecoverCompact   36           32           -11.11%
BenchmarkSigVerify        19           15           -21.05%

benchmark                 old bytes    new bytes    delta
-----------------------------------------------------------
BenchmarkRecoverCompact   1745         1617         -7.34%
BenchmarkSigVerify        977          848          -13.20%
```